Fix for images disappearing on tap
Summary:
Images were disappearing because we were marking them hidden on change, then waiting for TextKit to redo layout, but TextKit was not redoing layout at the right time.

Modified Lexical to specifically invalidate the layout for changed decorator nodes, which forces TextKit to redo the layout.

I also stopped marking the decorators as hidden ahead of layout -- with this fix, it was working, but in some conditions there was a bit of flicker where it quickly went hidden then visible. There should be minimal downside to this -- the TextKit layout process will handle marking the view as hidden if needed, and we're forcing this to happen every time now.

Reviewed By: thegreatercurve

Differential Revision: D44953924

fbshipit-source-id: 0783696bdc2f017e3c9b04661d5e2726d6431e4c