Fix for setting editor state
Summary:
We had an interesting bug here, which newly came to light with a recent refactor of how commands work. (To be clear, the bug was always there, but occurred in fewer situations previously.)

If the editor's pendingEditorState was overwritten (e.g. by calling `setEditorState()`) while inside an update block, then the editor's pendingEditorState would no longer match the activeEditorState attached to the current thread. This would cause problems, since things like `getLatest()` would pull from the `activeEditorState`.

Changed the logic so that if there's already a pendingEditorState, we modify the existing object, rather than changing it with a new object.

This logic is still a bit fragile, and is one of the oldest areas in Lexical iOS Core, so is probably worth a more holistic refactor at some point. Nontheless, this fixes the bug.

Reviewed By: fantactuka

Differential Revision: D45115630

fbshipit-source-id: eedf748544553705987be599ce41c4aa3354f4c6