Fixed undo coalescing
Summary:
Undo coalescing was broken if we sent updates that should be ignored. This was not previously noticed, but due to a recent refactor, we started sending many more such updates, which made the problem more noticeable.

(Undo coalescing is what makes it that you can press undo once to undo a whole word or sentence, rather than having to undo every single keystroke one at a time.)

Reviewed By: yannickl

Differential Revision: D45185049

fbshipit-source-id: c48aa2baf53a3809ef27933c97fe3e0d2459ebc6