Re-port insertText() from JS
Summary:
To accommodate changes in the web version of Lexical since we began the iOS project, as well as additional features now supported by iOS Lexical that were not present when we first wrote insertText(), I went back to the latest Lexical JS source and ported insertText() and deleteCharacter() line-by-line into Swift.

Practical problems this solves include some issues relating to formatting, e.g. whether formatting is applied to the selection upon certain actions (e.g. character deletion). It is also a good thing in general to have behaviour that is very similar to Lexical JS.

Note: as part of doing this work, I spotted a marked text bug: type some text, then apply a format (eg bold), then type a keystroke that initiates marked text. Instead of the expected character appearing, two non-marked characters appear after the caret. This bug is not new with this refactor -- it appears in the old codepath as well -- but I'm mentioning it here to document it.

Reviewed By: zurfyx

Differential Revision: D47930319

fbshipit-source-id: 67f743d2cc0a9b4c0f558fc7d48c7aa5050ed56a