Back out "Back out "[lexical-ios][PR] Re-port insertText() from JS""
Summary:
Original diff was D47930319, which was backed out in D48116843. Now re-applying the original diff with a fix.

D47930319 was removed due to it exposing a bug in `updateCaretSelectionForUnicodeCharacter()`.

I have confirmed that the best course of action is to remove `updateCaretSelectionForUnicodeCharacter()` entirely, as it is not needed in the iOS version due to the difference in how `modify()` works on the iOS platform (i.e. by calling an Apple-provided API to extend the selection by a character).

The bug in D47930319 was exposed by pressing backspace to delete a multi-codepoint UTF16 character, such as certain emoji. Tested in Pinecone, this version of the code does not suffer from this bug (emojis are all deleted correctly).

Reviewed By: zurfyx

Differential Revision: D48318727

fbshipit-source-id: 0127be1ae454b3b16dee4c4ac39fc823e7551266