Separate object to be text view delegate
Summary:
I was experiencing an infinite loop within UIKit when focussing a Lexical view for text input. It seems the issue was that Lexical's TextView is its own delegate. This diff fixes that.

However, I still have no idea why this is the first time I encountered the issue -- Lexical works fine on many other surfaces! There must be something specific to the integration I'm working on at the moment. Nontheless, this fix should be better all round.

Reviewed By: thegreatercurve

Differential Revision: D48397196

fbshipit-source-id: df81bcacf3565e9d1f9fc78276482ffd0c7773df