Height sizing
Summary:
Initial implementation of height sizing.

This is a fairly naive implementation so far, and I will hopefully improve it in the future. Notably, there is an occasional issue where the height is sized a fraction of a pixel too small, which causes the scroll view to temporarily allow scrolling. This will be fixed in a future diff.

Currently, I'm using component state to trigger a re-render. I don't think this is how `FBComposerTextView` does it, but I don't quite understand what they're doing to invalidate the layout!

Differential Revision: D48584597

fbshipit-source-id: eddc58d704da3544c3994351fd00c252346b1215