Made some bits of NativeSelection public
Summary:
The Lexical iOS philosophy is to do everything with the Lexical API where possible, and not to touch the UITextView underpinnings. However, since the Lexical API is not completely perfect, I'm coming round to the idea of exposing a few 'escape hatches' to access some of the UIKit stuff.

Note that it will still be recommended not to _modify_ any state within Lexical without going through a Lexical API. But in this case, I had a need to read the native selection in order to integrate with some legacy code that required selection ranges. Other people may have similar needs.

Differential Revision: D48950401

fbshipit-source-id: 3086159b1fba16bdb0be32ef07edab760e3caadf