Public access to text view
Summary:
I made it public to access the underlying UITextView inside a LexicalView.

I had initially wanted to avoid doing that. However I've now come around to it, since there will often be things that users of Lexical need to do that Lexical itself does not account for yet. In this case, I (as a user of Lexical) needed to use UITextView's `sizeThatFits:` method. I could (and probably should, in the future) modify LexicalView to support passing through the result of this. But for now I added an escape hatch.

Differential Revision: D48950402

fbshipit-source-id: 531b84c472961318e032808d95fd127e48b00fd9