Function to find string ranges from within Lexical using arbitrary predicates.
Summary:
There are many situations where it is useful to derive string ranges from Lexical EditorState. I'm currently working on a project where I talk to an API that expects string ranges for inline styles (e.g. bold/italic). I wrote a helper function that can help derive these.

Notably, I didn't assume an editor, so I couldn't read the values out of the range cache. I wanted the function to work inside an `editorState.read {}`.

The algorithm is pretty simple, doing a recursive tree traversal through the children of the start node. The caller provides a closure that says whether the node in question should be included in the ranges or not. The closure can also return `.inherit`, which will include the node if it is a child of an included node.

There is a version of the function that allows the closure to supply a payload, and the payloads are returned alongside the ranges.

The ranges are merged if they are contiguous. If the payload version of the function is being used, the merging only happens if the payloads are also equal.

Unit tests to come in a future diff.

Reviewed By: zurfyx

Differential Revision: D49610272

fbshipit-source-id: 26cacdffb1c4a08e23d157ccc7b6a66c6929f58c