Memory leak fixes (#40)
Summary:
This fixes 2/3 memory leaks in the Editor and makes the history plugin optional for the toolbar.

The third source of memory leaks is the History Plugin if anybody wants to take that on to fix.

This partially fixes issues referenced in: https://github.com/facebook/lexical-ios/issues/39

Pull Request resolved: https://github.com/facebook/lexical-ios/pull/40

Reviewed By: zurfyx

Differential Revision: D49729376

Pulled By: amyworrall

fbshipit-source-id: 6bf00712845affc66613106f6d9ce62d89d881e9