Fixed infinite loop in mentions
Summary:
Previously, I was trying to call setMode() inside the init method for a node. This caused an infinite loop, as the Lexical cloning logic thought the node needed cloning again (as it hadn't finished being cloned the first time!) when doing getWritable() inside setMode().

Temporarily fixed it by making `mode` public, so setting it directly during my init method; this is not a great long term solution however, as it could lead to new bugs if people don't realise they shouldn't access `mode` directly outside of initialisers. I will try and think of a better way this could be solved.

Differential Revision: D50363261

fbshipit-source-id: 23b9690bf50c7c890246b8719032470a62e3f6e1