LexicalMarkdownPlugin: say hello to markdown exports (#50)
Summary:
This is a first pass at implementing Markdown support for lexical.

It depends on Apple's swift-markdown package.

It needs tests and covering more edges cases (embedded lists, indentations, etc) - but is enough to get started with.

I am also working on a markdown import for our use case, which I will add to this plugin in the next day or so.

Pull Request resolved: https://github.com/facebook/lexical-ios/pull/50

Differential Revision: D51525285

Pulled By: amyworrall

fbshipit-source-id: 04065ac4fbf79c058e4285a511aa806fe9da8b59