Added facility to register commands that do not need editor update
Summary: An editor update is a costly operation, possibly taking up to 10ms. There are situations where your Lexical command's handler does not need to be inside an update block. A big example is logging. This diff allows command handlers to opt out of the auto-update-wrapper functionality. When applied to logging plugins, this can produce a drastic speedup.

Reviewed By: Fetz

Differential Revision: D53517219

fbshipit-source-id: ff525cf6010e37d3ed9fdcdc5f1e758a06d0243e