Fixed updates within update listeners
Summary:
Although it's not a recommended pattern to call update{} from within an update listener (see [docs](https://lexical.dev/docs/concepts/listeners) description of waterfall updates), it should still work, albeit inefficiently running the reconciler twice.

Previous to this diff, a bug in iOS Lexical meant that the reconciler would not be run after the second update, leaving the EditorState unreconciled. This diff fixes that, and correctly runs the reconciler after the second update.

Differential Revision: D54178279

fbshipit-source-id: 880d56eca74c6fb3e0b6fc879ad58a7d78e442d6