Correctly clean up lexical scope if exception is thrown
Summary:
There was a bug where if an exception was thrown within a Lexical update/read closure, the thread dictionary would be left with bad/old values on it, causing getActiveEditorState() to return an incorrect editor state during read operations. This is now fixed.

This bug became an issue due to my previous diff D54178279, which fixed another bug in update listeners that themselves trigger an update. I believe the previous issue was masking this issue until now.

Differential Revision: D54361984

fbshipit-source-id: ae3a0b5b9e79b4fe399eaea5515a4f331a77bce5