Use LexicalMarkdown to send markdown
Summary: Diff includes a change to LexicalMarkdown to wrap itself in an editor `read {}` block. This is a preferred way of building a Lexical API that takes an editor as a passed in parameter. (If the API expected to already be in a read block, it should get the editor using `getEditor()`. Since it doesn't, we'll do this.)

Reviewed By: yannickl

Differential Revision: D53044999

fbshipit-source-id: aefb401d8d6355f5859979c5723f5e09454137ec