Fixed read only check in update flow, fixes bug with child editors
Summary:
This fixes a bug when parent/child editors are used. (Currently, they are only used in the code for Tables.)

The issue related to how we track nested editor updates. When parent/child editors are used, a nested editor update may take place on a completely different editor. Thus, Lexical maintains a stack of nested editors in the thread dictionary, using `previousParentUpdateBlocksThreadDictionaryKey`.

This issue arose by performing a read only mode check in the wrong place. We were checking it when starting an update. However, the check done at that time was only checking the immediate parent update for read only mode. What we actually wanted to do was just ignore any read only 'updates' (i.e. read operations, or the implicit read that wraps listeners) entirely when considering nested updates. The goal of _that_ is so that we run the reconciler and perform the logic to move the pending editor state to the actual live editor state only at the outermost update for the current editor.

So, previously, the read only mode check was being done in the wrong place, and so the reconciler was run and the pending editor state migrated in a situation where we might still be in a nested update. Why does that matter? Because if the outer update had kept hold of any reference to the editor state, even implicitly by using `getActiveEditorState()` or a function that internally calls that, and used it after the nested update had been done but before the outer update had returned, then that code might end up working with an out-of-date copy of the editor state.

I fixed all of this by doing the read only check when _adding_ editors to the stack inside the thread dictionary, rather than when _reading_ them.

Just to be totally clear about how the read only check interfered with the nested editor check: in the previous code, if we used our incorrect logic for detecting read only mode, we might not set `isInsideNestedEditorBlock` to true. And if that was not set, then the full reconciler would be run, including the cleanup moving `pendingEditorState` to `editorState`, at the end of the inner nested update. And if _that_ happened, and the outer update had not finished its work, then the outer update would be acting on bad copies of the state that it would normally expect to rely on.

Differential Revision: D55430238

fbshipit-source-id: ebcba373c1386b324be244d6719cb452245159d0