Added Supernatural iOS app and dependencies
Summary:
allow-large-files

Added and buckified all dependencies under VendorLib, added and buckified the Supernatural (SCAP) source code, added aliases for buck and arc.

The 3P libs in VendorLib are private (using visibility in the BUCK file). The plan for these dependencies is to get them approved or replace them with Meta alternatives instead as soon as possible. Right now our highest priority is to get the code up and running on fbsource so we can do bug fixes and releases, migrate to our new Graph backend endpoints and start to replace the dependencies with Meta 1P alternatives, one by one.

`arc lint f` breaks our application, so we are investigating why these issues are occuring. We've added a number of lint exceptions for the more egregious errors, the remaining are notices or warnings, which will be addressed once our Github iOS repository is archived.

Reviewed By: jdifrancisco-supernatural

Differential Revision: D57969546

fbshipit-source-id: 0c355f4203c473d5ac27021139bbccf9434c9508