Adopt new method of UITextInputDelegate protocol
Summary:
`UITextInputDelegate` has a new required method in iOS 18.4 SDKs. Adopt the new
method, but mark it with `available` and also do not compile it when `UIKit` is
not new enough (in order to have access to `UIConversationContext`).

The implementation of the method follows the pattern of other methods in the
file.

Differential Revision: D73294690

fbshipit-source-id: 95456fb42f8a9b8ce71d2a0e8ed66c191b6d77c2